updater/unix.py: disable chroot for 0.4.2
The process in the jail cannot access encodings and cannot lookup
for FQDNs.  It seems not trivial to understand all the issues and
implications here, unfortunately.  So I have decided to disable
chroot for the next release cycle and to defer decision until I've
studied the problem a bit more.