neubot.nsi: do not autostart neubot
The problem is that when you autostart Neubot from the windows
installer it runs with too many privileges.  It seems not so
simple to deal with this issue and I have not time with a release
pending for the end of the week.  So, sorry about that, I will
disable this functionality until I find a proper fix.