database: fix regression with old clients
When an old client sends a dictionary with a missing field
the server will fail to commit.  And the result is not saved
into the database.

This commit adds _table_utils.py a wrapper function that takes
care of filling with None all the fields that are missing in
the passed dictionary.  The template is used as a reference for
the comparison.

This fixes a regression that prevents old Neubots to have their
results saved into the central database.  As well as to save them
into their private database.