table_config.py: byebye functional style
There were some functional code in database table management
but was not needed because always the consumer was a single
and simple function.  So, in the interest of keeping the tree
small, here's a commit that removes the last bit of functional
style.

As a bonus, this commit also removes one XXX from our tree.