net: Remove measurer
The measurer was mainly unused code.  So it makes sense to make
a little extra effort to thrash it, in the interest of keeping
the tree as small as possible.

The only one consumer of the measurer was the speedtest client,
which used it to calculare the average time to connect.  This
patch modifies StreamHandler to save into a vector the RTTs of
each connection.  So speedtest client can calculate the average
using that vector instead of the measurer.

In all other cases, the measurer was just a parameter passed
along and never used.