speedtest/client.py: fallback of measurer removal
The speedtest model is that there is a parent class that connects
and then it dispatches the work (negotiate, latency, download,
upload and collect) to child classes.  When I removed the measurer
I also modified the child collect class to use its self.rtt vector
to compute the average RTT.  Unfortunately, that vector is empty
because it's the parent that connects.  So the runtime raises a
ZeroDivisionError exception.  Avoid that by always overriding the
child rtts vector with the parent one.