net: simpler and more robust timeout model
This commit modifies the timeout model to make it simpler and
hopefully more robust.  The bottom line is that now there is
just a per-stream watchdog timeout.  This timeout is initialized
by default to 300 seconds and who needs to run forever should
override it.  The poller periodically invokes handle_periodic(),
which is a method fully-implemented by Pollable.  This method
returns True if the watchdog timeout is positive and a watchdog
timeout or more is passed since the Pollable was created.