net: Simplify task scheduling
No one is using unsched and resched facilities and no one is
passing arguments to scheduled functions.

So we can remove this features.

This is not harmful in the long run because I plan to replace
my task scheduling code with the standard library one, to shrink
even more the tree.  And because it's likely to be faster than
my code.