utils/blocks.py: new data generator
This new data generator is mostly compatible with the existing one,
neubot/blocks.py.  But the new generator has a startup time of about
1/100 of second on my netbook.  While the old one takes as long as
seven seconds.

The raw speed of the new data generator is comparable to the old
data generator one.  This because they use the same algorithm and
the difference is just that the new one uses real words, while
the old one generates words at random.

So we replace the old generator with the new one and we see what
happens.  If everything works OK, Neubot 0.4.3 will ship with the
new data generator.