debian: preinst: do not use curses to show license text
It turns out that using curses is more difficult than expected because
one has to consider the size of the terminal and make sure that the text
fits or take proper countermeasurements.  Additionally Federico pointed
out that I should show together both the human friendly and the full
privacy policy.

While it would be nice to wrestle with curses to accomplish both goals,
I think it is simpler to use less(1) to show the privacy policy to
the user.  I know that less(1) will be available because now I list
it as a preinstall dependency.

Additionally, this diff makes the script quiet by default and provides
a -v command line switch to make it verbose.  Since I do not use curses
anymore, I want the standard error to be quit when running in textual
mode, unless the user explicitly wants more information.

Finally, -C (for curses) is no longer appropriate and I've replaced it
with -T (for textual).