stream.py: stop passing exceptions upstream
If we unconditionally read exceptions from sys.exc_info() we
end up printing the last exception because the exception info
is never cleared.  We can modify the code to clear it each
time after usage, but, in the interest of keeping Neubot small,
I prefer to ditch the code that needs it.  We can live also
without those fancy prints that, by the way, were active only
in debugging mode.