privacy: Revert privacy policy to version 1.0
I've been busy with OWF11, data analysis and the AICA 2011 paper
for too much time.  And missed the expected 7 October deadline.
Now I want to roll out something soon.

The improved privacy policy is just a clarification for attorneys,
which can wait one more cycle.  When I implement it I also would
like to replace can_share with can_publish, for clarity.  And I have
more important things to do in this cycle.

This commit does not touches debian/control/preinst because it
needs to be nuked.  Unfortunately it doesn't work with Ubuntu
Software Center and possibly other noninteractive debconf frontends.

Discussed with and ok Federico.