debian: nuke python-based preinst
It is always a pity to get rid of good, simple and minimal code
but unfortunately Ubuntu Software Center runs the actual install
noninteractively and there is no way for this preinst to work
with it.  There is no DISPLAY where we can open the webkit window.
And there is no standard input where we can ask questions.  The
standard input is a pipe and when we ask questions and read from
it we deadlock the installer.

Let's start with reverting this piece of work and then I will
try to write something that can cope with the noninteractive
case as well.