debian: robust and least surprising reenable-repository algorithm
Ubuntu installer disables thirdparty repositories and appends a localized
message to the commentedout repository line.  To write robust code, I have
no choice but to look for a commentedout line followed by any comment.
Since that violates the rule of least surprise, add a warning and a link
to an explanation.

Please note that this should be OK with respect to e.g. Synaptic used
to disable the repository.  Because Synaptic just prepends the comment
sign in front of the repository line.

Discussed with and OK Federico.