debian: use cron.hourly for debugging purpose
I want this patch to stay intree for a release or two and
then I will move the script to cron.daily, which is the cron
category it really belongs to.

The patch has been tested with Ubuntu 11.04 in English and
Italian, simulating an upgrade.  In particular, the upgrade
procedure is started, so that the system disables Neubot
repository.  Then the commented out file is copied and the
upgrade is aborted.  On abortion the system restores the
statu quo ante.  At this point the copied commented out file
is copied over the restored one.  And the operator waits
for cron to run.  It worked in both cases.