debian: Privacy strategy that works (TM)
This commit finally implements a privacy strategy that works
on Debian and Ubuntu.  The first piece of the strategy is
that we check settings at the end of the installation and we
prompt the user via update-notifier if they are not OK.
The second piece of the strategy is that we run a new script
on gui login, which keeps annoying the user with libnotify
until settings are OK.

Neubot_webkit has been renamed neubot_gui and revamped to
fallback to webbrowser if WebKit and Gtk bindings aren't
available.  In addition, now the command does not attempt
to open WebKit window (or the browser) if DISPLAY isn't
available.  Finally, I've moved the neubot-not-running web
page code to neubot/www/not_running.html, because it is
not possible to load a string using webbrowser.

Neubot_notify is the new command that runs at gui login and
notifies the user if the privacy settings are not OK.  It does
not read the settings from the database but rather it queries
the Neubot daemon.  It is based on neubot_webkit code and does
not work if libnotify bindings are not available.  This is OK
for now, given that we mainly target Ubuntu users.  In the long
run it would be better to split Neubot into two packages and
list all the dependencies.

The postinst script has been extended and now it opens the
database and reads privacy settings.  If they are not OK it
writes a special file, which, in turn, causes update-notifier
to prompt the user to adjust privacy settings using Neubot
GUI.

While there this commit also modifies postinst to make it
reentrant.  Now it is possible to invoke it when Neubot is
already installed.  Because it does not fail if the user
and group already exist.  Needless to say, that's very handy
when you want to test the postinst script alone.

While there I've also modified the default shell of the
unprivileged user.  Now we use /usr/sbin/nologin instead
of /bin/false.