privacy: Bye bye special provisions
This commit removes special provisions that allow to perform
tests even if the user is not informed.  The changes affect
both the privacy policy and the actual code.

The privacy policy version number is bumped to 1.5 to signal
that a relevant change has happened.  Actually, the diff is
pretty small.  The commit just removes the text that describes
special provisions.  The affected files are: the text-only
privacy policy, the html privacy policy and its translations
in english and italian.

The bulk of the diff affects privacy.py.  First of all, this
commit adds functions that tell whether Neubot can run a test
and complain otherwise.  Now the rule is very simple: the user
must be informed and must have given the permission to save
her Internet address.  In addition, this commit simplifies the
checks performed when the user changes privacy settings from
the Web GUI.  The only accepted configuration is (1,1,*).
Finally, this commit rewrites privacy.py's main(), which now
allows a user to manage privacy settings from the command
line.  This is for users that run Neubot on a server.  BTW
this last change affects MacOSX postflight script, which used
privacy.py to finalize settings.

The remainder of the diff is composed of hunks that hook new
privacy checks to agent.py and rendezvous/client.py.  In
particular, now the rendezvous is not performed if privacy
settings are not OK.  The check is agent.py instead is just
to make some noise on the standard error.  So an admin may
notice that and adjust settings.

Note that it is still possible to write nonsense settings
into the database using the command line, which does not
perform checks.  The reason is that I don't want to be too
nazi: Someone that understands what is going on should be
able to hack into the settings quite easily.

While there, I've also taken the liberty of sorting imports
in agent.py and privacy.py.