win32: open browser if privacy not OK
This patch modifies rendezvous so that under NT it opens the
browser and points it to privacy.html when privacy settings
are not OK.  We can do that on NT because we run in the same
session of the user.  This is something annoying but I bet
it will be effective to prompt user to set permissions.  More
elaborate solution can always be stacked on top of this very
simple approach, BTW.

This patch also modifies the Web GUI to print a message that
explains why we popped up the browser.  The result is quite
ugly and I hope Alessio has some time to tweak the appearance
a bit.

While there, garbage collect an unused import.