neubot.nsi: install with user privilege level
Before this commit Neubot Windows installer used administrator
privileges.  This commit modifies neubot.nsi: now the installer
runs with the installing user privileges.

The main reason why I'm pushing this commit is that it allows
to run neubotw.exe at the end of the installation process.
That was the behavior of the installer, until I disabled that
when I realized neubotw.exe was executed with administrator
privileges.

In addition, the installation into a directory where the user
can write is DEFINITELY the way to go.  Especially given that
I want to implement Windows autoupdates as soon as possible.
And I don't want to take the security risk of doing that with
administrator privileges.

Another advantage of the unprivileged installer is that, if
privacy settings are not OK, neubotw.exe will open a browser
and ask the user to provide permissions.  And that happens
just at the end of the installation.  Of course it would be
better to handle privacy in the installer.  But, given the
debian experience, I'd rather roll out a simple solution NOW
and THEN I will do my best to write something fancier.

Unfortunately, the unprivileged installer does not come
without side effects.  One problem is that the installer
cannot uninstall old versions of Neubot.  Because they were
installed with administrator privileges.  For this reason
I've implemented a message box, which requests to uninstall
manually.

Another problem is that the installation directory is the
user's profile dir.  This is a bit more surprising than the
usual 'C:\Program Files\neubot' but I cannot find a least
surprising alternative at the moment.

While making the installer unprivileged, I've taken the
liberty to remove the amount of crap we install.  Now there
are just two symlinks: one to autostart on login and the
other to open the Web GUI.  And I've modified the file to
use 'Neubot' (i.e. using the uppercase N) because uppercase
names are more common in Windows.