http/server.py: honour the connection header
Modify http server stream send_response behavior to honour the connection
header.  If the request contains the indication that the connection
should be closed or the protocol was HTTP/1.0, set the 'connection:
close' header in the response.  If the response contains the indication
that the connection should be closed, close it.

Also remove a stream.close() in api/server.py because it is redundant
after this change.