debian/neubot_notify: die when graphic session is over
When graphic session is over the session dbus is not reachable anymore
and pynotify fails.  Rewrite the code so that when it fails we do not
catch the exception.  This guarantees that no "detached" neubot_notify
will linger around.

Unfortunately, we cannot do the same for the case when we open the
webbrowser.  And the pending release does not leave time to investigate
the problem more in depth.  So I've decided to remove the webbrowser code,
which means the script either runs under Ubuntu or (likely) raises an
import error.  I'm sorry.

While there rewrite the code that checks privacy settings to account
for the string representation of booleans.  This should be fixed as
of now, but some extra robustness will never hurt.