neubot_notify: tune sleep interval and log exceptions
It is not optimal to keep polling Neubot every 15 seconds when
we know settings are OK.  An occasional poll every hour will do.

Log and reraise pynotify exceptions because I've noticed that
the script died after several hours on my Ubuntu 11.10 but I
don't know why because the exception was lost.  Hope that will
happen again so I understand better in what cases this piece
of code can fail.