debian: postinst: avoid races between neubot and sqlite3
Neubot may lock the database and sqlite3 may do the same, which
means it is not wise to start Neubot and then check the database
using sqlite3.  Better to check the database first and THEN do
the usual invocation of Neubot.  Should fix an installation error
seen in the new ubuntu virtual machine.