UNIX: always install neubot_notify and neubot_gui
Reflecting on what is installed when you type `make install` I
noticed that neubot_notify and neubot_gui should both be installed
to complement neubot.  They are part of the typical Neubot
experience.  For this reason this patch modifies the Makefile.
Note that I've not moved the two extra bins from debian/ because
the release is pending.  I will do that after release.

In addition this patch modifies neubot_notify to produce some
noticeable effect even when pynotify is not installed.  This
is quirky.  Again, I will deal with this issue after the pending
0.4.3 release of Neubot.  The plan is to allow for two pkgs,
one with all the stuff and one with just the -nox part.