UNIX: simplify the task of creating a package
Simplify the task of creating a package.  Now the packager can
tweak the result by overriding BINDIR, DATADIR, SYSCONFDIR, MANDIR
and LOCALSTATEDIR.  Default settings are unaffected.

Now the Makefile installs the notifier and the viewer for every
platform, not just for Debian.  They are always installed and
the packager decides what dependencies to add.  If dependencies
are not met, the notifier and the viewer will print an error
and exit.

Now the notifier and the viewer are internal commands and the
Makefile installs just one command in BINDIR: `neubot`.  The
reason is that our idea of Neubot is a bundle that includes the
notifier and the viewer.  Having separate commands for certain
features clearly does not encourage bundling.

Technically, this patch moves all the UNIX-only stuff in the
UNIX directory.  In turn, this directory is organized to allow
for simple Makefile code.  Basically, the `_install` target
of Makefile has been rewritten and I have also propagated the
changes.

The patch is crafted to minimize the changes and the only
relevant difference with before is that start-neubot-daemon
is no longer installed.  Because of that, I had to delete
some lines from the manual page.  Another extra modification
is that I have added a very simple unit test.