icons: remove the folder and its content
While working on a Win32 patch I noticed that the only consumer
of neubot.ico was setup.py.  In turn, neubot.ico was same-as
favicon.ico.  So, I've changed setup.py to use the latter.

Now, it turns out that not only neubot.ico is unused, but all
the files in icons are not used.  Since they can be easilly
generated from UNIX/share/icons/.../neubot.svg, I think there
is no point in keeping icons around.