Debian: simplify packaging
When I wrote Makefile I didn't know about Debian packaging and
the result was that there were too many targets.  This can be
simplified a bit.  There should be no functional change, except
that some targets have now been merged.

Apart from Makefile changes, this patch replaces changelog with
a compressed changelog.  And garbage collects the special-purpose
neubot executable with hardcoded /usr/bin/python.  If the system
is so screwed that the Python picked up by env does not work with
Neubot, probably it is better to reinstall the system.