BUGS: it may be confusing to mention the ML address like this
Many people (including me) use a first-fit strategy and is likely
to stop as soon as they find something that matches their criteria.
To discourage people with this attitude (including me) do not put
the mailing list address in the BUGS file, so that they need to click
on the weblinks and read the whole story.

Of course, the website have been changed accordingly.