Makefile: honour $(LOCALSTATEDIR)
Create the directory and make sure we replace the @LOCALSTATEDIR@
placeholder with the actual value.

We used to create the directory before the partial rewrite of the
install command and we should do the same now.  Spotted because the
non-standard-dir error from lintian was now missing.

The second change, to replace the pattern, is preparatory work
because I plan to move the database to /var/lib/neubot on GNU/Linux
systems.  In turn, this will remove the last lintian error and I
can try to ask Debian folks to add the package.