regress/Makefile/install.sh: more checks
Make sure that we don't install something with @PLACEHOLDERS@ by
mistake.

While there, change the name of the know-good installation signature
because I look forward to add more tests for Debian.

In addition, sort the know-good installation signature and sort the list of
installed files, because different systems will find in different order.