Debian: get dependency right
We need python, sqlite3 and procps to be installed before Neubot
is installed.  Because they are needed by postinst and postrm.

In addition, Neubot depends on cron and bsdutils to keep up to date
its thirdparty repository.  This dependency will go away when we will
enter into debian.

Finally, python-webkit is required by the viewer component and
python-notify by the notifier component.  I add them because my
opinion is that the typical Neubot experience includes both.