Debian: support for -nox flavour
I don't know whether there is someone out there that uses Neubot
in headless mode.  If there is, I want to support that folk.

This patch is a bit hacking and I will try to improve this support
in the future.  For now, what is important is that the patch is intree
and that I will create the package.