Makefile: remember to compile modules on plain install
Remember to compile modules on plain install so that we do not
only install .py files but we also install .pyc files.

Likewise, compile modules when we simulate an install to generate
the list of the files to uninstall.

The combination of this two changes fixes uninstalling Neubot
from a system where I installed from sources and actually I
had used Neubot.  The reason why uninstall was partially failing
before this patch is that there were around the .pyc files and
the uninstall script was not expecting that.

Note that this patch assumes that the user has installed a
command named python.  However, I know that OpenBSD does not
enforce a system python by default and installs python2.7.
For situations like that one, the Makefile uses a PYHTON variable
so the user (who knows better) can always override.