bittorrent: simplify and clarify maximum message length
First of all, rename SMALLMESSAGE to MAXMESSAGE, given that the
code to process "jumbo" messages has been removed.

Additionally, remove an unneeded check in peer.py, because it
suffices to check the length in stream.py.

Moreover, in stream.py, perform the check right after we have
read the length and not while reading the body.

Finally, raise the maximum message length to be double the size
of the pieces we usually send and receive.