scripts/debugmem: allow to monitor objects count
Add a command line switch -c, which instructs the script to
monitor the number of objects tracked by Python GC.

Also add a -1 command line switch, which instructs the script to
query the Neubot daemon state just once.  Often, I just want to
see the number of tracked objects, not to enter into a loop that
monitors such number.