speedtest/server.py: keep memory fragmentation under control
Do not read incoming requests body because this leads to
memory fragmentation, especially when testing with very fast
clients.  Indeed, when we receive a huge HTTP request and we
don't ignore it, we need to allocate bytes for such request.
With this patch, instead, the server memory usage is under
control again (after more than one hour of testing).

This was the real cause behind the memory usage, however I'm
happy with the new memory instrumentation I've added to Neubot,
which for sure will help to keep the server under control.
Thanks to all the folks that discussed with me the supposed
memory leak: Roberto D'Auria, Marco Scopesi, Enrico Calzavara
and Elias Carotti.