regress: add unit test Debian/control/postinst
This commit adds a test unit for Debian/control/postinst, which
tries to ensure that the script works as expected under a number
of operating conditions.  Some bits can be improved but for the
upcoming release this will be enough.  For example it would be
nice to check strictly new user and group attributes.