regress: Debian/reinst: account for notification
The first time we install the database does not exist and so the
installer creates a Window asking the user to update privacy settings.
So, give the user more time to close the Window (checking whether it
works is not the purpose of this test).

Note that this test actually fails and I have a patch that fixes
the postinst script and shows the alert when the database does not
exist because it's the first install.