rendezvous: more strict rules for redirection
Redirect to other servers only if the user provided all the
privacy permissions.  My understanding is that, without this
patch, M-Lab folks may not accept Neubot on their platform.

At the moment I'm waiting for clarification, but my guess
is that Thomas is right.  Following the guess, I'm doing what
I expect he will ask me to do in advance :-).

If I'm wrong, it'll be easy to revert this patch and return
to 0.4.4 statu quo.