MacOS: tell the login screen not to pick Neubot user
As reported by many, when you install Neubot on MacOSX the new
user `Neubot` appears in the login screen.

The fix is to set its password to the literal *, so the login
screen knows that the user is disabled and does not pick it up.
For more info on that topic, please see::

	http://superuser.com/questions/70156

This patch actually does something more than setting the user
password.  Specifically, it:

1. makes sure that the code that sets/updates real user name,
   password and shell is run both when the user is created and
   during updates;

2. switches the shell from /sbin/nologin to /usr/bin/false,
   which is what is done e.g. for the _ssh user;

3. provides a more meaningful real user name, clarifying that
   _neubot and _neubot_update are for privilege separation.

This (and in particular 1) should fix the login screen problem
for users that have already installed Neubot.  I prefer this
solution than telling them to go fix the problem at hand using
the command line.