M-Lab: scripts to deploy, start and stop
This commit adds scripts to deploy Neubot to M-Lab, start and stop
the installed copy of Neubot.

The script to deploy is a five step-process:

        - The first step stops the installed copy of Neubot (if
          any) and it removes the installed files (if any).  This
          step must be run as root because Neubot on M-Lab runs on
          behalf of its pseudouser ``_neubot``.

        - The second step copies Neubot tarball and current revision's
          hash to the server.

        - The third step unpacks the tarball and pycompiles the
          *.py files inside.

        - The fourth step ensures that all the prerequisites for
          running Neubot are met and then starts it.  Again, this
          must be run with root privileges.

        - The fifth step gets rid of the tarball.

The script to stop neubot reads the PID from the pidfile and then
kills the PID.  Since server.py was not writing the pidfile, I have
added the code to do that (just one line).

The script that checks for prerequisites makes sure that we have
the _neubot user and group.  Additionally, it installs a modified
/etc/rc.local that should start Neubot.

The script to start Neubot also makes sure that syslog is running,
because I want to be able to read logs.

The last remark is that the deploy script starts neubot invoking
/etc/rc.local instead of the startup script directly.  The idea
is to pass for the same code path, regardless of the fact that I
am deploying Neubot or the server is restarted.