M-Lab/ip_addr.sh: get slivers IP address
An M-Lab node contains many slivers and each sliver has its own
Internet address.  Given the domain name of an M-Lab node that
belongs to our slice, this script returns the IPv4 address of our
sliver in such node.

We need such IPv4 address because our master server must redirect
to that address and not to the address of the node.

I'm sure there are better ways to do that, but this solution is
acceptable for now.