M-Lab/servers.py: write approxymate location of nodes
Basically I want to know the country and the continent where a
node is.  With this information, I can setup a static table that
redirects users to the closest set of nodes.

This script does not use the Internet address of a node, because
what we need is the Internet address of the sliver in the node.
And this is done by another script.

The best way to know the location of a node is probably via
PlanetLab API.  Being a bit in a hurry, I've decided to roll
out this solution instead.  Albeit approxymate, this solution
is IMHO enough, given that I need coarse grained geoloc.