M-Lab/redir_table.py: code to build redirection table
This scripts builds the redirection table using the following
strategy:

1. when there is at least one server in the same country, choose
   one of those servers at random;

2. when there is at least one server in the same continent, pick
   one of those servers at random;

3. when there are no servers in the continent, pretend the user
   lives in another continent (always the same, of course).

So, for example, DE users are redirected to one of the few
servers in Germany.  SE users get one of the many servers
deployed across Europe.  African users are always redirected
to European servers.

From the point of view of research, case 2. and 3. may not be
bad, because, at the end of the day, the Neubot will probe a
number of different network paths.  Which likely will provide
more information, when we analyze the results.  So, this may
become a value for us.

What wonders me is the point of view of the user.  It may
happen that the measured goodput after test N is lower than
after test (N-1) just because Neubot has been redirected
over a longer path.  For example, you live in Sweden, test
N-1 was performed against Germany and test N is performed
against Greece.  This may weaken sensibly one important
feature of Neubot, i.e. that it is valuable for users
because it provides information.

For this reason, I commit this patch but I do not enable
the redirection immediately.  I clearly need to reflect
a bit more on this topic and seek for better alternatives
(for example DONAR DNS in least-latency mode here comes
to my mind).