M-Lab/countries.dat: allow comments and convert to UTF-8
To allow comments we need to modify redir_table.py to check for
and honour comments.

Once we have support for comments, we can move countries.txt
content at the beginning of countries.dat as a comment and we
can delete countries.txt.

While there, convert countries.dat content from ISO-8859 to
UTF-8 (it changes just a couple of lines).