Toplevel: Update to privacy policy 2.0
This commit updates the toplevel directory to version 2.0 of
the privacy policy.  There are just three files where the word
"privacy" is mentioned in the toplevel directory: ChangeLog,
PRIVACY and README.  It turns out that we just need to update
the PRIVACY file, at this level, replacing the old with the
new privacy policy.

=== NEW PRIVACY POLICY ===
Version 2.0 of the privacy policy contains the following set
of changes:

1. Neubot will not work unless the user has provided all the
   privacy permissions;

2. we have renamed `can_publish` the `can_share` variable.

Change 1. is because M-Lab allows to use its server platform
only when all privacy permissions are provided, and we want to
force all users to use M-Lab, given the tremendous advantages
of using it.

Change 2. is because `can_publish` explains more clearly
what is the purpose of the permission (`can_share` has been
criticized because it was too vague and not precise).
=== NEW PRIVACY POLICY ===

I would like to thank Eleonora Bassi, who provided insightful
comments on the previous incarnation of the policy, and helped
me to draft an earlier text of this new version.  Feedback &
suggestions also from Claudio and Federico.

After this commit the word "share" does not show up anymore
in PRIVACY.