neubot/www: Update to privacy policy 2.0
This commit updates neubot/www directory to version 2.0 of the
privacy policy.

=== NEW PRIVACY POLICY ===
Version 2.0 of the privacy policy contains the following set
of changes:

1. Neubot will not work unless the user has provided all the
   privacy permissions;

2. we have renamed `can_publish` the `can_share` variable.

Change 1. is because M-Lab allows to use its server platform
only when all privacy permissions are provided, and we want to
force all users to use M-Lab, given the tremendous advantages
of using it.

Change 2. is because `can_publish` explains more clearly
what is the purpose of the permission (`can_share` has been
criticized because it was too vague and not precise).
=== NEW PRIVACY POLICY ===

Here is a list of all the files that contained the word
"privacy" along with what was changed:

- neubot/www/css/style.css
    - Nothing

- neubot/www/header.html
    - Nothing

- neubot/www/js/privacy.js
    - Replace can_share with can_publish (honour 2)
    - Unhide privacy warning if can_publish is False (honour 1)

- neubot/www/lang/en.js
    - Regen

- neubot/www/lang/it.js
    - Add my copyright
    - Translate the new policy
    - Say "salvare" instead of "collezionare"
    - Replace "condividere" with "pubblicare" (honour 2)
    - Honour 1 in the bottom-page warning
    - Cosmetic changes

- neubot/www/privacy.html
    - Replace the old with the new privacy policy
    - Remove wrong "i18n i18n_privacy_settings_2" class [*]
    - Say clearly that we collect and publish "for research purposes"
    - Replace "share" with "publish" (honour 2)
    - Honour 1 in the bottom-page warning

[*] The parser that generates lang/en.js stops when it finds the
first occurrence of the i18n class, i.e. it does not support nesting
of i18n classes.  In privacy.html "i18n i18n_privacy_settings_2"
was the outer class that contained nested classes.  But the objects
of translation were the inner classes not the outer one.  So, I
have removed the outer class, for the script to correctly put the
inner classes in ``lang/en.js``.

After this commit, the word "share" does not show up anymore
below ``neubot/www``.