Debian: Update to privacy policy 2.0
This commit brings the Debian subdirectory up to date
with v2.0 of the privacy policy.

This commit modifies ``Debian/control/postinst`` to use
neubot.privacy API (and specifically the -t command line
option) to check whether it needs to prompt the user to
update its privacy settings.  After this change, we don't
need anymore the sqlite3 dependency, so I'm removing it
from control files.

Changes tested with the regression test scripts, which,
by the way, have been modified to cope with the new privacy
policy.