neubot/log.py: listify(): don't unconditionally connect to database
This commit is just for the sake of correctness.  We must not connect
to the database in listify() if the database is not enabled.  It is much
more robust to return an empty list instead.

Incidentally, this commit also fixes neubot/log.py regression test,
when such test is run on a system where the _neubot user does not
exist.  In such case, DATABASE.connection() raises a SystemExit error,
trying to fix privacy permissions.  This is part of the oddities of
the database code and, sooner or later, I will have to deal with
that.