database: Update to privacy policy 2.0
This commit updates the database subsystem to use the new
privacy policy, which basically means replacing ``can_share``
with ``can_publish``, because we're at the storage layer.

Of course, we also need to crank the version number and
write a migrator that updates the existing schemas and
data.  This time the migrator may be slow, because it has
to rewrite the bittorrent and speedtest table, using the
code committed a few commits ago for the purpose of helping
this diff.

Tested migrating a 16MiB database on my MacOSX system,
using the new shiny main() added to migrate.py to allow
for simple testing.  It takes about 3.0 seconds to do
all the stuff, which is not good because this laptop is
one of my fastest computers.  We'll see how that will
work on the netbook.

NB: From this commit on, collect_allowed() will not accept
anymore an object and expects a dictionary.