clients servers: Update to privacy policy 2.0
The BitTorrent client has been modified to use the new variable
name ``privacy_can_publish``, while speedtest and rendezvous clients
stay with the old name ``privacy_can_share``, because it's simpler
than modifying the message classes to deal with both names (yes
we need to support both names on server side).

All servers have been modified so that they can work with both the
old and the new name, even if BitTorrent is the only one that is
actually going to see the new name.

Tested running clients and server pairs, making sure that there is
backwards compatibility and that: (i) rendezvous server does not
redirect to M-Lab unless you have all permissions; (ii) rendezvous
client does not even talk to the server unless you have all the
permissions; (iii) speedtest and bittorrent server does not save
the result when they're not allowed to do that.

The difference in behavior between (i) and (iii) depends on the fact
that rendezvous server uses count_valid() while bittorrent and speedtest
servers use collect_allowed(), which is less strict on purpose, to
guarantee (iii) for a while for backward compatibility.

I think backward compatibility will not be a problem because nonone
will be able to use M-Lab unless she explicitly invokes Neubot in
nonstandard ways or modifies the code, which both void the warranty.

NB: I updated the way collect_allowed() is invoked by speedtest client
because that function does not accept an object anymore.