table_speedtest.py: Move glue to speedtest/client.py
This commit moves the code that glues XML and JSON representations
from table_speedtest.py to speedtest/client.py, where it really
belongs, because the table code has now become format agnostic and
no other piece of code is referencing this glue.  So, move it to
the file that really needs it, which is a preliminary step before
decommissioning.

Tested running a speedtest test vs. master.neubot.org on the command
line and verifying that the result is really written into the data-
base.