database/_table_utils.py: allow to rename columns
I need this piece of code because soon I'll need to be able
to s/can_share/can_publish/ in bittorrent and speedtest table
of the database, when I'll upgrade the database subsystem to
version 2.0 of the privacy policy.

This is not straightforward because sqlite3 does not allow to
do that via a simple 'ALTER TABLE'.  The strategy to do that
is described on stackowerflow.com at this URI:

    http://stackoverflow.com/questions/805363

Discussed with Roberto.

Tested using the attached regression test.